Create Sum of odd and even numbers
Write a program to calculate the sum of odd numbers and the sum of even numbers separately up to a given integer N. The program should return both sums: one for odd numbers and the other for even numbers, separated by a space.

Here, we will be using a for loop and if-else to calculate the output. The numbers divisible by 2 are added to even, and the other numbers are added to odd.