Create Sum of Squares of Natural Numbers
Write a program to calculate the sum of the squares of the first N natural numbers. The program should return the sum of squares from 1 to N.

Here, Taking user input as n, and initializing sum as 0 (starting). A for loop is used where i is starting from 1 (Natural numbers do start from 1) and calculating sum as sum=sum+i*i where i is squared and added to the sum.