Create Absolute Value
Write a program to find the absolute value of a given integer I. The absolute value of a number is the distance between that number and zero on the number line, without considering its sign. If the number is positive or zero, the absolute value is the same as the number. If the number is negative, the absolute value is the positive equivalent of that number.

Here, we used if else by comparing the n value to zero whether its greater or smaller.