Create 2D Array
Given a 2D array, an hourglass is a subset of values with indices falling in the following pattern:

a b c  
  d  
e f g

There are hourglasses in an array. This is the sum of the values in an hourglass. Calculate the hourglass sum for every hourglass in , then print the hourglass sum.

Example
-9 -9 -9  1 1 1 
 0 -9  0  4 3 2
-9 -9 -9  1 2 3
 0  0  8  6 6 0
 0  0  0 -2 0 0
 0  0  1  2 4 0
The  hourglass sums are:

-63, -34, -9, 12, 
-10,   0, 28, 23, 
-27, -11, -2, 10, 
  9,  17, 25, 18
The highest hourglass sum is from the hourglass beginning at row , column :