Create Sparse Array using Hashing- Unordered Map
Hashing is a technique used to map data to a fixed-size value called a hash code, usually to store or find it fast. 

It is of syntax unordered_map<string,int>mp, where the string address or key value is stored in integer type and the name of the map. 

We will read the string into the map. And increasing the count of the freq of the string. We will read query and print the freq[query]  if it matches and prints the count.